Turn speed is now based off of ship movement and ship can no longer move backwards

+++ Level/Assets/Scripts/Ship/shipCameraControl.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class shipCameraControl : MonoBehaviour
 {
-    [SerializeField] public int sensHort;
+    [SerializeField] public float sensHort;
     [SerializeField] public int sensVert;
 
     [SerializeField] bool invert;
